+++ uniffi_bindgen/src/bindings/ruby/gen_ruby/mod.rs
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use anyhow::Result;
-use askama::Template;
+use rinja::Template;
 
 use heck::{ToShoutySnakeCase, ToSnakeCase, ToUpperCamelCase};
 use serde::{Deserialize, Serialize};
@@ -106,15 +106,10 @@ impl Config {
 pub struct RubyWrapper<'a> {
     config: Config,
     ci: &'a ComponentInterface,
-    canonical_name: &'a dyn Fn(&Type) -> String,
 }
 impl<'a> RubyWrapper<'a> {
     pub fn new(config: Config, ci: &'a ComponentInterface) -> Self {
-        Self {
-            config,
-            ci,
-            canonical_name: &canonical_name,
-        }
+        Self { config, ci }
     }
 }
 
@@ -122,7 +117,7 @@ mod filters {
     use super::*;
     pub use crate::backend::filters::*;
 
-    pub fn type_ffi(type_: &FfiType) -> Result<String, askama::Error> {
+    pub fn type_ffi(type_: &FfiType) -> Result<String, rinja::Error> {
         Ok(match type_ {
             FfiType::Int8 => ":int8".to_string(),
             FfiType::UInt8 => ":uint8".to_string(),
@@ -152,7 +147,7 @@ mod filters {
         })
     }
 
-    pub fn literal_rb(literal: &Literal) -> Result<String, askama::Error> {
+    pub fn literal_rb(literal: &Literal) -> Result<String, rinja::Error> {
         Ok(match literal {
             Literal::Boolean(v) => {
                 if *v {
@@ -188,26 +183,32 @@ mod filters {
         })
     }
 
-    pub fn class_name_rb(nm: &str) -> Result<String, askama::Error> {
+    pub fn class_name_rb(nm: &str) -> Result<String, rinja::Error> {
         Ok(nm.to_string().to_upper_camel_case())
     }
 
-    pub fn fn_name_rb(nm: &str) -> Result<String, askama::Error> {
+    pub fn fn_name_rb(nm: &str) -> Result<String, rinja::Error> {
         Ok(nm.to_string().to_snake_case())
     }
 
-    pub fn var_name_rb(nm: &str) -> Result<String, askama::Error> {
+    pub fn var_name_rb(nm: &str) -> Result<String, rinja::Error> {
         let nm = nm.to_string();
         let prefix = if is_reserved_word(&nm) { "_" } else { "" };
 
         Ok(format!("{prefix}{}", nm.to_snake_case()))
     }
 
-    pub fn enum_name_rb(nm: &str) -> Result<String, askama::Error> {
+    pub fn enum_name_rb(nm: &str) -> Result<String, rinja::Error> {
         Ok(nm.to_string().to_shouty_snake_case())
     }
 
-    pub fn coerce_rb(nm: &str, ns: &str, type_: &Type) -> Result<String, askama::Error> {
+    pub fn coerce_rb<S1: AsRef<str>, S2: AsRef<str>>(
+        nm: S1,
+        ns: S2,
+        type_: &Type,
+    ) -> Result<String, rinja::Error> {
+        let nm = nm.as_ref();
+        let ns = ns.as_ref();
         Ok(match type_ {
             Type::Int8 => format!("{ns}::uniffi_in_range({nm}, \"i8\", -2**7, 2**7)"),
             Type::Int16 => format!("{ns}::uniffi_in_range({nm}, \"i16\", -2**15, 2**15)"),
@@ -252,7 +253,8 @@ mod filters {
         })
     }
 
-    pub fn check_lower_rb(nm: &str, type_: &Type) -> Result<String, askama::Error> {
+    pub fn check_lower_rb<S: AsRef<str>>(nm: S, type_: &Type) -> Result<String, rinja::Error> {
+        let nm = nm.as_ref();
         Ok(match type_ {
             Type::Object { name, .. } => {
                 format!("({}.uniffi_check_lower {nm})", class_name_rb(name)?)
@@ -270,7 +272,7 @@ mod filters {
         })
     }
 
-    pub fn lower_rb(nm: &str, type_: &Type) -> Result<String, askama::Error> {
+    pub fn lower_rb(nm: &str, type_: &Type) -> Result<String, rinja::Error> {
         Ok(match type_ {
             Type::Int8
             | Type::UInt8
@@ -305,7 +307,7 @@ mod filters {
         })
     }
 
-    pub fn lift_rb(nm: &str, type_: &Type) -> Result<String, askama::Error> {
+    pub fn lift_rb(nm: &str, type_: &Type) -> Result<String, rinja::Error> {
         Ok(match type_ {
             Type::Int8
             | Type::UInt8