+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 [All changes in [[UnreleasedUniFFIVersion]]](https://github.com/mozilla/uniffi-rs/compare/v0.28.2...HEAD).
 
+### What's changed?
+
+- Switching jinja template engine from askama to rinja.
+
 ## v0.28.2 (backend crates: v0.28.2) - (_2024-10-08_)
 
 ### What's new?
@@ -31,8 +35,6 @@
 - `uniffi.toml` of crates without a `lib` type where ignored in 0.28.1
 - Python: Fixed a bug when enum/error names were not proper camel case (HTMLError instead of HtmlError).
 - Python: Fixed the class hierarcy generated for traits ((#2264)[https://github.com/mozilla/uniffi-rs/issues/2264])
-
-[All changes in v0.28.2](https://github.com/mozilla/uniffi-rs/compare/v0.28.1...v0.28.2).
 
 ## v0.28.1 (backend crates: v0.28.1) - (_2024-08-09_)